Rename app_row
