+++ appRow/proguard-rules.pro
