+++ appRow/src/main/AndroidManifest.xml
