+++ app_drawer/build.gradle
@@ -8,5 +8,5 @@ android {
 
 dependencies {
     implementation project(path: ":apps")
-    implementation project(path: ":app_row")
+    implementation project(path: ":appRow")
 }
