+++ settings.gradle
@@ -44,6 +44,6 @@ include ":app_search"
 include ":app_drawer"
 include ":launcher"
 include ":apps"
-include ":app_row"
+include ":appRow"
 include ":wordle"
 include ":camera"
